fix(agents): tighten multi-runtime recovery lifecycle
Rollback left runtime state live. Every sweep expired runtime rows even
with the flag off, so after disabling multi-runtime mode the rows workers
had stopped refreshing were retired and their agents marked offline while
they were serving traffic perfectly well. Expiry is now inert unless the
mode is on.

A crashed runtime kept its session. Expiry retired the runtime but left the
session it owned, which still looked to orphan recovery like a live
process, so the task it held only came back through the much slower
stalled-task path. Retirement now removes that runtime's sessions in the
same transaction; a live sibling's sessions are untouched, and the freed
task reaches the normal recovery paths.

Assignment ran before expiry. `codeLevelTriage` auto-assigned pool tasks
and only afterwards retired dead runtimes, so a task could be handed to an
agent the same sweep was about to mark offline — with nothing left to poll
for it. Expiry now runs first, and auto-assignment additionally skips
agents whose runtime rows are all dead.

Retired runtimes accumulated. Runtime identity is per boot, so keeping
rows meant one per boot per agent forever; nothing reads a runtime once it
stops being live, so they are pruned once past the liveness window.

Resetting the task limit forced every agent to one slot, which left a lead
unable to start the follow-up its own task waits on. Reset now returns the
agent to its role default.

Feature-specific persistence moves out of db.ts into src/be/multi-runtime.ts;
the env/flag helper in src/utils stays DB-free so worker-side code can keep
importing it. Drops an unused runtime-list helper along the way.
