+++ src/be/db.ts
@@ -69,8 +69,6 @@ import type {
   ReasoningEffort,
   RepoGuidelines,
   RoutingAffinity,
-  RuntimeInstance,
-  RuntimeInstanceStatus,
   ScheduledTask,
   ScheduledTaskSummary,
   ScriptRun,
@@ -868,92 +866,6 @@ export function updateAgentMaxTasks(id: string, maxTasks: number): Agent | null 
     )
     .get(maxTasks, id);
   return row ? rowToAgent(row) : null;
-}
-
-/**
- * Agent-scoped swarm_config key holding the logical maxTasks policy in
- * multi-runtime mode. Operators edit this row; `agents.maxTasks` stays the
- * enforcement mirror that hasCapacity()/getRemainingCapacity() read. Same
- * column-vs-config split as HARNESS_PROVIDER.
- */
-export const AGENT_MAX_TASKS_CONFIG_KEY = "AGENT_MAX_TASKS";
-
-/** Bounds match AgentSchema.maxTasks, so a mirrored write stays schema-valid. */
-function parseAgentMaxTasksValue(raw: string): number | null {
-  const str = raw.trim();
-  if (!/^\d+$/.test(str)) return null;
-  const value = Number(str);
-  return value >= 1 && value <= 100 ? value : null;
-}
-
-export function getAgentMaxTasksConfig(agentId: string): SwarmConfig | null {
-  return (
-    getSwarmConfigs({ scope: "agent", scopeId: agentId, key: AGENT_MAX_TASKS_CONFIG_KEY })[0] ??
-    null
-  );
-}
-
-/**
- * Reconcile the AGENT_MAX_TASKS policy row with its agents.maxTasks mirror,
- * never consuming the runtime-reported concurrency.
- *
- * With no row yet, seed from the persisted agents.maxTasks — the value the
- * swarm is already enforcing — so enabling multi-runtime mode cannot change
- * behaviour. An existing row is authoritative and repairs the mirror instead.
- *
- * Runs inside the registration transaction so concurrent runtimes serialize:
- * the first seeds, the rest observe.
- */
-export function reconcileAgentMaxTasksPolicy(agentId: string): void {
-  const agent = getAgentById(agentId);
-  if (!agent) return;
-  const existing = getAgentMaxTasksConfig(agentId);
-  if (existing) {
-    const authoritative = parseAgentMaxTasksValue(existing.value);
-    if (authoritative !== null && authoritative !== (agent.maxTasks ?? 1)) {
-      updateAgentMaxTasks(agentId, authoritative);
-    }
-    return;
-  }
-  upsertSwarmConfig({
-    scope: "agent",
-    scopeId: agentId,
-    key: AGENT_MAX_TASKS_CONFIG_KEY,
-    value: String(agent.maxTasks ?? 1),
-    description:
-      "Logical agent max-tasks policy (seeded from agents.maxTasks at first multi-runtime registration)",
-  });
-}
-
-/**
- * Config upsert that keeps the agents.maxTasks enforcement mirror in step
- * with an agent-scoped AGENT_MAX_TASKS write, in one transaction, so policy
- * and enforcement cannot diverge — including with no runtime connected.
- * Every other key passes straight through.
- */
-/**
- * Reset the enforcement mirror when the policy row is deleted. Without this
- * the deleted limit survives on `agents.maxTasks` and the next registration
- * re-seeds the policy from it, making the delete a no-op.
- */
-export function resetAgentMaxTasksMirror(agentId: string): void {
-  updateAgentMaxTasks(agentId, 1);
-}
-
-export function upsertSwarmConfigWithPolicyMirror(
-  data: Parameters<typeof upsertSwarmConfig>[0],
-): SwarmConfig {
-  const scopeId = data.scope === "agent" ? data.scopeId : null;
-  if (!scopeId || data.key !== AGENT_MAX_TASKS_CONFIG_KEY) {
-    return upsertSwarmConfig(data);
-  }
-  return getDb().transaction(() => {
-    const row = upsertSwarmConfig(data);
-    // Guard, not validation — both write paths already ran validateConfigValue.
-    const value = parseAgentMaxTasksValue(data.value);
-    if (value !== null) updateAgentMaxTasks(scopeId, value);
-    return row;
-  })();
 }
 
 export function updateAgentProvider(id: string, provider: ProviderName): Agent | null {
@@ -8736,32 +8648,6 @@ export function cleanupAgentSessions(agentId: string): number {
   return result.changes;
 }
 
-/**
- * Startup cleanup for one runtime: drop this agent's sessions that no live
- * runtime owns, which is what a restarting process leaves behind (its
- * previous boot used a different instance id).
- *
- * A live sibling's session must survive — deleting it would strand its task
- * with no session row, and the orphan sweep would then requeue work that
- * process is still executing.
- */
-export function cleanupRuntimeSessions(agentId: string): number {
-  const result = getDb()
-    .prepare(
-      `DELETE FROM active_sessions
-       WHERE agentId = ?
-         AND (
-           runtimeInstanceId IS NULL
-           OR runtimeInstanceId NOT IN (
-             SELECT id FROM runtime_instances
-             WHERE agent_id = ? AND status = 'active' AND last_seen_at >= ?
-           )
-         )`,
-    )
-    .run(agentId, agentId, runtimeLivenessCutoff());
-  return result.changes;
-}
-
 /** Update providerSessionId on an active session identified by taskId */
 export function updateActiveSessionProviderSessionId(
   taskId: string,
@@ -8796,192 +8682,6 @@ export function reassociateSessionLogs(runnerSessionId: string, realTaskId: stri
     .prepare("UPDATE session_logs SET taskId = ? WHERE sessionId = ? AND taskId != ?")
     .run(realTaskId, runnerSessionId, realTaskId);
   return result.changes;
-}
-
-// ============================================================================
-// Runtime Instance Functions (multi-runtime worker tracking)
-// ============================================================================
-
-interface RuntimeInstanceRow {
-  id: string;
-  agent_id: string;
-  status: string;
-  reported_slots: number;
-  metadata: string | null;
-  last_seen_at: string;
-  created_at: string;
-  updated_at: string;
-  created_by: string | null;
-  updated_by: string | null;
-}
-
-function rowToRuntimeInstance(row: RuntimeInstanceRow): RuntimeInstance {
-  let metadata: Record<string, unknown> | null = null;
-  if (row.metadata) {
-    try {
-      metadata = JSON.parse(row.metadata) as Record<string, unknown>;
-    } catch {
-      metadata = null;
-    }
-  }
-  return {
-    id: row.id,
-    agentId: row.agent_id,
-    status: row.status as RuntimeInstanceStatus,
-    reportedSlots: row.reported_slots,
-    metadata,
-    lastSeenAt: row.last_seen_at,
-    createdAt: row.created_at,
-    updatedAt: row.updated_at,
-  };
-}
-
-/**
- * Register (or refresh) a runtime instance serving a logical agent. A
- * re-registration from the same process refreshes its row in place rather
- * than accumulating rows.
- *
- * The `DO UPDATE ... WHERE` guard makes runtime ownership permanent: an id
- * already held by one agent can never be reassigned to another, so a
- * duplicate id cannot move a live runtime between agents. Returns null in
- * that case (no row written).
- */
-export function upsertRuntimeInstance(instance: {
-  id: string;
-  agentId: string;
-  reportedSlots: number;
-  metadata?: Record<string, unknown> | null;
-}): RuntimeInstance | null {
-  const metadataJson = instance.metadata ? JSON.stringify(instance.metadata) : null;
-  const row = getDb()
-    .prepare<RuntimeInstanceRow, [string, string, number, string | null]>(
-      `INSERT INTO runtime_instances (id, agent_id, status, reported_slots, metadata)
-       VALUES (?, ?, 'active', ?, ?)
-       ON CONFLICT(id) DO UPDATE SET
-         status = 'active',
-         reported_slots = excluded.reported_slots,
-         metadata = COALESCE(excluded.metadata, runtime_instances.metadata),
-         last_seen_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now'),
-         updated_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now')
-       WHERE runtime_instances.agent_id = excluded.agent_id
-       RETURNING *`,
-    )
-    .get(instance.id, instance.agentId, instance.reportedSlots, metadataJson);
-  return row ? rowToRuntimeInstance(row) : null;
-}
-
-export function getRuntimeInstanceById(id: string): RuntimeInstance | null {
-  const row = getDb()
-    .prepare<RuntimeInstanceRow, [string]>("SELECT * FROM runtime_instances WHERE id = ?")
-    .get(id);
-  return row ? rowToRuntimeInstance(row) : null;
-}
-
-export function getRuntimeInstancesForAgent(agentId: string): RuntimeInstance[] {
-  return getDb()
-    .prepare<RuntimeInstanceRow, [string]>(
-      "SELECT * FROM runtime_instances WHERE agent_id = ? ORDER BY created_at ASC, id ASC",
-    )
-    .all(agentId)
-    .map(rowToRuntimeInstance);
-}
-
-/**
- * Refresh a live runtime instance's liveness from the worker's ping cadence.
- *
- * Matches only an `active` row owned by the pinging agent, so a ping can
- * neither resurrect a closed runtime nor reach another agent's row.
- * Registration stays the only writer of `active`. Returns false when nothing
- * matched, which callers use as the signal that the identity is not live.
- */
-export function touchRuntimeInstance(id: string, agentId: string): boolean {
-  const result = getDb()
-    .prepare(
-      `UPDATE runtime_instances
-       SET last_seen_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now'),
-           updated_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now')
-       WHERE id = ? AND agent_id = ? AND status = 'active'`,
-    )
-    .run(id, agentId);
-  return result.changes > 0;
-}
-
-/** Mark one runtime instance offline (scoped to its agent, like touch). */
-export function markRuntimeInstanceOffline(id: string, agentId: string): RuntimeInstance | null {
-  const row = getDb()
-    .prepare<RuntimeInstanceRow, [string, string]>(
-      `UPDATE runtime_instances
-       SET status = 'offline', updated_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now')
-       WHERE id = ? AND agent_id = ? RETURNING *`,
-    )
-    .get(id, agentId);
-  return row ? rowToRuntimeInstance(row) : null;
-}
-
-/**
- * Minutes without a ping before a runtime stops counting as live. The runner
- * pings every poll iteration (~2s), so this is a wide margin that will not
- * expire a healthy idle worker; it matches the "worker clearly dead" window
- * the stalled-task classifier already uses.
- */
-export function runtimeStaleThresholdMinutes(): number {
-  return Number(process.env.RUNTIME_STALE_THRESHOLD_MIN) || 5;
-}
-
-function runtimeLivenessCutoff(): string {
-  return new Date(Date.now() - runtimeStaleThresholdMinutes() * 60 * 1000).toISOString();
-}
-
-/**
- * Runtimes currently serving an agent. A crashed process never reaches
- * `/close`, so `active` alone would keep a dead runtime counted forever;
- * liveness is therefore the conjunction of status and a fresh `last_seen_at`.
- */
-export function countActiveRuntimeInstancesForAgent(agentId: string): number {
-  const result = getDb()
-    .prepare<{ count: number }, [string, string]>(
-      `SELECT COUNT(*) as count FROM runtime_instances
-       WHERE agent_id = ? AND status = 'active' AND last_seen_at >= ?`,
-    )
-    .get(agentId, runtimeLivenessCutoff());
-  return result?.count ?? 0;
-}
-
-/**
- * Retire runtimes that stopped pinging and take their agents offline when
- * nothing live remains. Runs in the heartbeat sweep; the per-agent
- * reconciliation shares one transaction with the expiry so an agent is never
- * observed with zero live runtimes but an available status.
- */
-export function expireStaleRuntimeInstances(): { expired: number; agentsOffline: number } {
-  return getDb().transaction(() => {
-    const cutoff = runtimeLivenessCutoff();
-    const stale = getDb()
-      .prepare<{ agent_id: string }, [string]>(
-        `SELECT DISTINCT agent_id FROM runtime_instances
-         WHERE status = 'active' AND last_seen_at < ?`,
-      )
-      .all(cutoff);
-    if (stale.length === 0) return { expired: 0, agentsOffline: 0 };
-
-    const result = getDb()
-      .prepare(
-        `UPDATE runtime_instances
-         SET status = 'offline', updated_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now')
-         WHERE status = 'active' AND last_seen_at < ?`,
-      )
-      .run(cutoff);
-
-    let agentsOffline = 0;
-    for (const { agent_id } of stale) {
-      if (countActiveRuntimeInstancesForAgent(agent_id) > 0) continue;
-      const agent = getAgentById(agent_id);
-      if (!agent || agent.status === "offline") continue;
-      updateAgentStatus(agent_id, "offline");
-      agentsOffline++;
-    }
-    return { expired: result.changes, agentsOffline };
-  })();
 }
 
 // ============================================================================
